refactor(agent-status): isolate task inventory parsing
