+++ src/shared/claude-subagent-roster.ts
@@ -1,18 +1,15 @@
 import { AGENT_STATUS_MAX_SUBAGENTS, type AgentSubagentSnapshot } from './agent-status-types'
+import type { ClaudeBackgroundAgentTask } from './claude-background-task-inventory'
+
+export {
+  hasActiveClaudeNonAgentBackgroundWork,
+  readClaudeBackgroundAgentTasks
+} from './claude-background-task-inventory'
 
 /** Mirrors the wire-normalization id cap in agent-status-types. Enforced at
  *  upsert so an over-long id can't gate the pane 'working' while being
  *  invisible in the emitted snapshots (which drop such ids). */
 const CLAUDE_SUBAGENT_ID_MAX_LENGTH = 64
-const CLAUDE_TERMINAL_BACKGROUND_TASK_STATUSES = new Set([
-  'idle',
-  'completed',
-  'failed',
-  'killed',
-  'cancelled',
-  'canceled',
-  'timed_out'
-])
 
 /** Live subagents/teammates tracked for one Claude pane, keyed by the
  *  provider-assigned `agent_id` from SubagentStart/SubagentStop payloads.
@@ -58,19 +55,6 @@ export type TrackedClaudeSubagent = {
    *  teammate-shaped. Never cleared: the listing mode of an id can't change
    *  mid-life. */
   listedAsSubagentTask?: true
-}
-
-/** One agent entry from the `background_tasks` array Claude attaches to Stop
- *  (and SubagentStop) hook payloads. Non-agent tasks do not become rows. */
-export type ClaudeBackgroundAgentTask = {
-  id: string
-  agentType?: string
-  description?: string
-  running: boolean
-  /** True for `type: "teammate"` entries. Their ids never match lifecycle
-   *  agent_ids and they report "running" permanently — even after the named
-   *  agent finished — so they carry no per-agent state at all. */
-  teammate: boolean
 }
 
 /** Agent-team/named-agent lifecycle ids are `a<name>-<hex>` while one-shot
@@ -148,87 +132,6 @@ export function stopClaudeSubagent(roster: ClaudeSubagentRoster, id: string): vo
     return
   }
   tracked.state = 'idle'
-}
-
-/** Shell tasks and session crons outlive a lead Stop but do not belong in the agent roster. */
-export function hasActiveClaudeNonAgentBackgroundWork(
-  hookPayload: Record<string, unknown>
-): boolean {
-  const sessionCrons = hookPayload['session_crons']
-  if (Array.isArray(sessionCrons) && sessionCrons.length > 0) {
-    return true
-  }
-  const backgroundTasks = hookPayload['background_tasks']
-  return (
-    Array.isArray(backgroundTasks) &&
-    backgroundTasks.some((item) => {
-      if (typeof item !== 'object' || item === null) {
-        return false
-      }
-      const task = item as Record<string, unknown>
-      return task.status === 'running' && task.type !== 'subagent' && task.type !== 'teammate'
-    })
-  )
-}
-
-/** Read the agent-typed entries of a hook payload's `background_tasks` field.
- *  `present: false` means the field was absent/malformed (older Claude builds),
- *  so callers must keep their tracked roster instead of clearing it. */
-export function readClaudeBackgroundAgentTasks(hookPayload: Record<string, unknown>): {
-  present: boolean
-  tasks: ClaudeBackgroundAgentTask[]
-  truncated: boolean
-  hasRunningNonAgentTask: boolean
-} {
-  const raw = hookPayload['background_tasks']
-  if (!Array.isArray(raw)) {
-    return { present: false, tasks: [], truncated: false, hasRunningNonAgentTask: false }
-  }
-  const tasks: ClaudeBackgroundAgentTask[] = []
-  let truncated = false
-  let hasRunningNonAgentTask = false
-  for (const item of raw) {
-    if (typeof item !== 'object' || item === null) {
-      truncated = true
-      continue
-    }
-    const obj = item as Record<string, unknown>
-    const taskType = typeof obj.type === 'string' ? obj.type.trim().toLowerCase() : ''
-    const taskStatus = typeof obj.status === 'string' ? obj.status.trim().toLowerCase() : ''
-    const isTerminal =
-      taskStatus.length > 0 && CLAUDE_TERMINAL_BACKGROUND_TASK_STATUSES.has(taskStatus)
-    if (taskType.length === 0) {
-      truncated = true
-      hasRunningNonAgentTask ||= !isTerminal
-      continue
-    }
-    const isAgentTask = taskType === 'subagent' || taskType === 'teammate'
-    // Why: future non-agent types and nonterminal labels must fail active; only typed agent rows or explicit terminal states can safely retire work.
-    if (!isAgentTask && !isTerminal) {
-      hasRunningNonAgentTask = true
-    }
-    if (!isAgentTask) {
-      continue
-    }
-    if (typeof obj.id !== 'string' || obj.id.trim().length === 0) {
-      truncated = true
-      continue
-    }
-    if (tasks.length >= AGENT_STATUS_MAX_SUBAGENTS) {
-      // Why: a capped inventory cannot prove a tracked id is absent; callers
-      // must retain unlisted rows rather than deleting live overflow tasks.
-      truncated = true
-      continue
-    }
-    tasks.push({
-      id: obj.id.trim(),
-      agentType: typeof obj.agent_type === 'string' ? obj.agent_type : undefined,
-      description: typeof obj.description === 'string' ? obj.description : undefined,
-      running: !isTerminal,
-      teammate: taskType === 'teammate'
-    })
-  }
-  return { present: true, tasks, truncated, hasRunningNonAgentTask }
 }
 
 /** Fold a lead Stop's `background_tasks` into the lifecycle-tracked roster.
